# chronos_app/packages/api-documentation/src/yml/agent-protocol.openapi.yaml
openapi: 3.1.0
info:
    title: Chronos Agent Protocol
    version: 1.6.0
    summary: Wire-format contract for the Chronos Agent Registry and MCP Gateway.
    description: |
        Two-direction protocol between the Chronos control plane and external HTTP
        agents.

        - **Platform → Agent.** What an external HTTP agent must implement to be
          invokable via the Chronos Agent Registry. Health probe + an OpenAI-compatible
          chat completions endpoint with two Chronos extension fields.
        - **Agent → Platform.** What Chronos exposes for tool brokering. Agents call
          back into the MCP Gateway to invoke registered MCP tools without holding
          MCP-server credentials directly.

        See `docs/agent-registry.md` and `docs/mcp-gateway.md` for the operator
        guides. v1.6 supports the `bearer` and `header` outbound-auth shapes;
        OAuth2 client-credentials is a follow-up. v1.6 supports `streamable-http`
        and `sse` MCP transports; `stdio` is reserved for v1.8.
    license:
        name: Apache 2.0
        identifier: Apache-2.0
        url: https://github.com/popularowl/chronos/blob/main/LICENSE.md
    contact:
        name: Chronos
        email: saul@popularowl.com
servers:
    - url: http://localhost:8001
      description: External HTTP agent (illustrative — your service)
    - url: http://localhost:3000/api/v1
      description: Chronos control plane (default dev)
...
security: []
paths:
...
    /v1/chat/completions:
        post:
            tags:
                - agent-runtime
            summary: OpenAI-compatible chat completions with Chronos extensions
            description: |
                The Chronos dispatcher forwards invocations as POSTs to
                `<serviceEndpoint>/v1/chat/completions`. The body is an
                OpenAI chat-completions request augmented with two Chronos
                extension fields; the same values are also sent as headers.

                For tool use, call back into Chronos using
                `x_chronos_mcp_gateway_url` (or the `x-chronos-mcp-gateway-url`
                header) with `Authorization: Bearer <agent.mcpGatewayToken>`.
                See `/api/v1/mcp-gateway/{agentId}/tools/invoke`.

                If the request body has `stream: true`, return an SSE stream
                in OpenAI's chunk format; the dispatcher pipes it back to the
                client chunk-by-chunk. Otherwise return a JSON
                ChatCompletion object.
            operationId: agentChatCompletions
            parameters:
                - in: header
                  name: Authorization
                  required: false
                  schema:
                      type: string
                  description: |
                      Whatever `outboundAuth` resolved to. For
                      `{ type: 'bearer', … }` this is `Bearer <token>`. For
                      `{ type: 'header', name: '<X>', … }` the credential is
                      injected as that named header instead.
                - in: header
                  name: x-chronos-call-id
                  required: true
                  schema:
                      type: string
                      format: uuid
                  description: |
                      Same value as `x_chronos_call_id` in the body. Echo
                      back via the callback's `callId` field.
                - in: header
                  name: x-chronos-mcp-gateway-url
                  required: true
                  schema:
                      type: string
                      format: uri
                  description: Same value as `x_chronos_mcp_gateway_url` in the body.
                - in: header
                  name: accept
                  required: false
                  schema:
                      type: string
                  description: |
                      `text/event-stream` when the request body has
                      `stream: true`, else `application/json`.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentChatCompletionsRequest'
            responses:
                '200':
                    description: Successful completion (JSON or SSE).
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ChatCompletion'
                        text/event-stream:
                            schema:
                                type: string
                                description: |
                                    OpenAI-format SSE stream. Each chunk is a
                                    `data: { ... }\n\n` line; terminate with
                                    `data: [DONE]`.
                '400':
                    $ref: '#/components/responses/AgentBadRequest'
                '401':
                    description: Auth presented to your agent did not pass.
                '500':
                    description: Internal agent error. Chronos surfaces this as 502 to its caller.
...
components:
    securitySchemes:
        mcpGatewayToken:
            type: http
            scheme: bearer
            bearerFormat: hex
            description: |
                Per-agent MCP gateway token. Stored on the `Agent` row as
                `mcpGatewayToken`. Rotatable from the Agent detail UI; rotation
                is single-step (no overlap window in v1.6).
    schemas:
        AgentChatCompletionsRequest:
            type: object
            description: |
                OpenAI chat-completions request body augmented with Chronos
                extension fields. The standard OpenAI fields (`model`,
                `messages`, `stream`, `temperature`, etc.) are forwarded
                unchanged.
            required:
                - messages
                - x_chronos_mcp_gateway_url
                - x_chronos_call_id
            additionalProperties: true
            properties:
                model:
                    type: string
                    description: Echoed by Chronos; agents may ignore.
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/ChatMessage'
                stream:
                    type: boolean
                    default: false
                x_chronos_mcp_gateway_url:
                    type: string
                    format: uri
                    description: |
                        Absolute URL of the MCP gateway endpoint this agent
                        should call back into for tool use, e.g.
                        `https://chronos.example.com/api/v1/mcp-gateway/<agentId>/tools/invoke`.
                        Same value also sent as the `x-chronos-mcp-gateway-url`
                        request header.
                x_chronos_call_id:
                    type: string
                    format: uuid
                    description: |
                        Per-invocation correlation ID. Echo back as `callId`
                        in any tool-callback request so audit lines join up.
        ChatMessage:
            type: object
            required:
                - role
                - content
            properties:
                role:
                    type: string
                    enum: [system, user, assistant, tool]
                content:
                    type: string
                tool_call_id:
                    type: string
                tool_calls:
                    type: array
                    items:
                        type: object
                        additionalProperties: true
        ChatCompletion:
            type: object
            description: Standard OpenAI ChatCompletion response.
            additionalProperties: true
            required:
                - id
                - object
                - created
                - choices
            properties:
                id:
                    type: string
                object:
                    type: string
                    const: chat.completion
                created:
                    type: integer
                model:
                    type: string
                choices:
                    type: array
                    items:
                        type: object
                        properties:
                            index:
                                type: integer
                            message:
                                $ref: '#/components/schemas/ChatMessage'
                            finish_reason:
                                type: string
                                enum: [stop, length, tool_calls, content_filter]
                usage:
                    type: object
                    properties:
                        prompt_tokens:
                            type: integer
                        completion_tokens:
                            type: integer
                        total_tokens:
                            type: integer
...
    responses:
        AgentBadRequest:
            description: Malformed inbound request to the agent.
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            error:
                                type: string